pyproject.toml
[tool.poetry]
name = "heipy"
version = "0.1.0"
description = ""
authors = ["hwisang.jeon <hwisang.jeon@navercorp.com>"]
readme = "README.md"
include = [{ path = "typings/**", format = ["sdist"] }]

[tool.poetry.dependencies]
python = "^3.12"
jpype1 = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
